pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "aind-mri-targeting"
description = "Tools to plan and execute MRI-guided targeting experiments"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
  { name = "Allen Institute for Neural Dynamics" },
  { name = "Galen Lynch", email = "galen@galenlynch.com" },
  { name = "Yoni Browning", email = "yoni.browning@alleninstitute.org" },
]
classifiers = ["Programming Language :: Python :: 3"]
readme = "README.md"
dynamic = ["version"]

dependencies = [
  'aind-anatomical-utils>=0.1,<0.2',
  'aind_mri_utils>=0.2.1,<2',
  'numpy>=1.20,<3',
  'pynrrd>=1,<2',
  'SimpleITK>=2,<3',
  'trimesh[recommend]>=4,<5',
]

[project.urls]
Repository = "https://github.com/AllenNeuralDynamics/aind-mri-targeting/"
Issues = "https://github.com/AllenNeuralDynamics/aind-mri-targeting/issues"
Changelog = "https://github.com/AllenNeuralDynamics/aind-mri-targeting/releases"

[project.scripts]
headframe_centers_of_mass = "aind_mri_targeting.scripts.headframe_centers_of_mass:main"
install_notebooks = "aind_mri_targeting.scripts.install_notebooks:main"
calculate_headframe_transforms = "aind_mri_targeting.scripts.calculate_headframe_transforms:main"
fit_implant = "aind_mri_targeting.scripts.fit_implant:main"

[project.optional-dependencies]
dev = ['codespell', 'pytest', 'pytest-cov', 'ruff', 'interrogate']
docs = [
  'mkdocs',
  'mkdocs-material',
  'mkdocstrings[python]',
  'pymdown-extensions',
  'ruff',
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = { attr = "aind_mri_targeting.__version__" }
readme = { file = "README.md", content-type = "text/markdown" }

[tool.ruff]
line-length = 120
target-version = 'py311'

[tool.ruff.lint]
extend-select = ['Q', 'RUF100', 'C90', 'I']
extend-ignore = []
mccabe = { max-complexity = 14 }
pydocstyle = { convention = 'numpy' }

[tool.codespell]
skip = '.git,*.pdf,*.svg,./bonsai,*.bonsai,./docs/_build,uv.lock'
ignore-words-list = 'nd'

[tool.pytest.ini_options]
addopts = "--strict-markers --tb=short --cov=src --cov-report=term-missing --cov-fail-under=70"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.interrogate]
ignore-init-method = true
ignore-magic = true
ignore_module = true
fail-under = 30
exclude = [
  "__init__.py",
  "tests",
  "docs",
  "build",
  "setup.py",
  "examples",
  "site",
]
